move airdrop to own test
